refactor: pull reserve logic out of sections.ts
