+++ packages/translator/src/visitors/placeholder.ts
@@ -3,12 +3,8 @@ import { isNativeTag } from "@marko/babel-utils";
 import { isOutputHTML } from "../util/marko-config";
 import { callRuntime, getHTMLRuntime } from "../util/runtime";
 import evaluate from "../util/evaluate";
-import {
-  ReserveType,
-  getSection,
-  reserveScope,
-  getSectionId,
-} from "../util/sections";
+import { getSection } from "../util/sections";
+import { ReserveType, reserveScope } from "../util/reserve";
 import * as writer from "../util/writer";
 import * as walks from "../util/walks";
